Refactor the code structure and add two files vmx.c and ept.c for Initing VMX-ENV and Memory-Virtualization.
